backend/requirements.txt
torch
diffusers
transformers
accelerate
peft
onnxruntime-gpu
pillow
flask
flask-cors
numpy